runeseal/requirements.txt
typer[all]
requests
rich
python-dotenv
